create-edit-albums

+++ app/controllers/songs_controller.rb
@@ -23,7 +23,7 @@ class SongsController < ApplicationController
     @song.user = current_user
     if @song.save
       flash[:success] = t ".add_song_successfully"
-      redirect_to users_my_song_url
+      redirect_to my_songs_url
     else
       flash[:danger] = t ".add_song_fail"
       render :new
@@ -33,9 +33,9 @@ class SongsController < ApplicationController
   def edit; end
 
   def update
-    if @song.update_attributes song_params
+    if @song.update song_params
       flash[:success] = t ".song_updated"
-      redirect_to users_my_song_url
+      redirect_to my_songs_url
     else
       flash[:danger] = t ".song_update_failed"
       render :edit
@@ -43,12 +43,16 @@ class SongsController < ApplicationController
   end
 
   def destroy
-    if @song.destroy
-      flash[:success] = t ".delete_successfully"
+    if check_current_obj_user? @song
+      if @song.destroy
+        flash[:success] = t ".delete_successfully"
+      else
+        flash[:danger] = t ".delete_fail"
+      end
     else
-      flash[:danger] = t ".delete_fail"
+      flash[:danger] = t ".must_current_user"
     end
-    redirect_to users_my_song_url
+    redirect_to my_songs_url
   end
 
   private
